updated property and label query mechanism query properties now produces instance-property mapping

+++ JPS_Chatbot/UI/data_preparation/Wiki_basic_info/query_species_classes_and_labels.py
@@ -1,5 +1,8 @@
 from util.SPARQL_Query_Wiki import SPARQL_Query_for_Wiki
 import json, re, time
+from os import listdir
+from os.path import isfile, join
+
 
 SPARQL_template = '''
 PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#> 
@@ -20,50 +23,61 @@ query_wiki = SPARQL_Query_for_Wiki()
 
 distinct_classes = []
 
-with open('query_log', 'w') as f:
-    f.write(str(time.time()) + '\n')
-    f.close()
+# with open('query_log', 'w') as f:
+    # f.write(str(time.time()) + '\n')
+    # f.close()
 
-counter = 0
+
+mypath = './instance_info'
+existing_files = [f for f in listdir(mypath) if isfile(join(mypath, f)) and f.startswith('Q')]
+
+
+
+counter = len(existing_files)
 
 start_time = time.time()
 with open('WIKI_URI_LIST') as f:
     instances = json.loads(f.read())
     print('number of instances', len(instances))
     for instance in instances:
-        counter = counter + 1
-        print('iterated', counter, 'out of', len(instances))
-        print('already took', round(time.time() - start_time, 2))
         id = re.search(r'Q[0-9]+', instance)[0]
-        SPARQL_query = SPARQL_template % (id, id)
-        try:
-            results = query_wiki.get_results(SPARQL_query)
-            with open('instance_info/%s' % id, 'w') as f:
-                f.write(json.dumps(results, indent=4))
-                f.close()
-            bindings = results['results']['bindings']
-            for b in bindings:
-                if 'class' in b:
-                    class_uri = b['class']['value']
-                    if 'classLabel' in b:
-                        class_label = b['classLabel']['value']
-                
-                        if class_uri in distinct_classes: # it already exists 
-                            # then do nothing, it is repeated 
-                            pass
-                        else:
-                            tmp = {class_uri: class_label}
-                            distinct_classes.append(tmp)
-                            with open('distinct_classes', 'w') as f:
-                                f.write(json.dumps(distinct_classes))
-                                f.close()
-                     
-        except:
-            with open('query_log', 'a') as f:
-                f.write(instance + '\n')
-                f.close()
-            pass        
-        time.sleep(1)
+        if id in existing_files:
+            pass
+        
+        else:
+    
+            counter = counter + 1
+            print('iterated', counter, 'out of', len(instances))
+            print('already took', round(time.time() - start_time, 2))
+            SPARQL_query = SPARQL_template % (id, id)
+            try:
+                results = query_wiki.get_results(SPARQL_query)
+                with open('instance_info/%s' % id, 'w') as f:
+                    f.write(json.dumps(results, indent=4))
+                    f.close()
+                bindings = results['results']['bindings']
+                for b in bindings:
+                    if 'class' in b:
+                        class_uri = b['class']['value']
+                        if 'classLabel' in b:
+                            class_label = b['classLabel']['value']
+                    
+                            if class_uri in distinct_classes: # it already exists 
+                                # then do nothing, it is repeated 
+                                pass
+                            else:
+                                tmp = {class_uri: class_label}
+                                distinct_classes.append(tmp)
+                                with open('distinct_classes', 'w') as f:
+                                    f.write(json.dumps(distinct_classes))
+                                    f.close()
+                         
+            except:
+                with open('query_log', 'a') as f:
+                    f.write(instance + '\n')
+                    f.close()
+                pass        
+            time.sleep(0.1)
